# shared/schemas/api.openapi.yaml
openapi: 3.0.3
info:
  title: Family Trivia API
  description: Backend API for the Family Trivia game
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local SAM development
paths:
...
  /create-custom-category:
    post:
      summary: Create a custom category
      operationId: createCustomCategory
      tags: [Categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, categoryName]
              properties:
                userId:
                  type: string
                categoryName:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Created category
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  category:
                    $ref: "#/components/schemas/CustomCategory"
...
  /celebrations:
    post:
      summary: Create a celebration
      operationId: createCelebration
      tags: [Celebrations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, name, date, addedBy]
              properties:
                type:
                  type: string
                  enum: [birthday, anniversary, holiday, other]
                name:
                  type: string
                date:
                  type: string
                description:
                  type: string
                addedBy:
                  type: string
                recurring:
                  type: boolean
                year:
                  type: integer
      responses:
        "200":
          description: Created celebration
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  celebration:
                    $ref: "#/components/schemas/Celebration"
    get:
      summary: Get all celebrations
      operationId: getCelebrations
      tags: [Celebrations]
      responses:
        "200":
          description: Celebrations list
          content:
            application/json:
              schema:
                type: object
                properties:
                  celebrations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Celebration"
                  lastUpdated:
                    type: string
    delete:
      summary: Delete a celebration
      operationId: deleteCelebration
      tags: [Celebrations]
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
...
components:
  schemas:
...
    CustomCategory:
      type: object
      required: [id, title, description, createdBy, createdAt, type]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
        type:
          type: string
          enum: [custom]
...
    Celebration:
      type: object
      required: [id, type, name, date, addedBy, addedAt, recurring]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [birthday, anniversary, holiday, other]
        name:
          type: string
        date:
          type: string
        description:
          type: string
        addedBy:
          type: string
        addedAt:
          type: string
        recurring:
          type: boolean
...